service-provider/requirements.txt
Flask==3.0.0
google-generativeai==0.8.3
python-dotenv==1.0.0
flask-cors==4.0.0